Add notmuch path support
muchsync already supports adding a specific path to the muchsync binary,
this add similar support for notmuch. It's possible that this could be
done better, but I'm not a professional C++ developer, so I don't know
all the various libraries and best practices.

Jetbrains added some extra stuff to the directory while I was trying to
debug, so I figured I'd add that stuff to the ignore in case anyone else
ended up using clion to work on this project.